requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-extensions==3.1.1
django-redis==4.12.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
flake8==3.9.0
flake8-django==1.1.1
flake8-import-order==0.18.1
flake8-print==4.0.0
kombu==5.0.2
mccabe==0.6.1
Pillow==8.1.2
pkg-resources==0.0.0
prompt-toolkit==3.0.17
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.0
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
